Moved the State object definition into the ra_sched file
